refactor: Update imports to use '@/lib/utils' instead of '@/lib' across multiple components; adjust Vite config aliasing for improved clarity
