+++ react-app/src/components/ui/checkbox.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox"
 import { CheckIcon } from "lucide-react"
 
-import { cn } from "@/lib"
+import { cn } from "@/lib/utils"
 
 function Checkbox({
   className,
