+++ react-app/src/components/ui/dropdown-menu.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
 import { CheckIcon, ChevronRightIcon, CircleIcon } from "lucide-react"
 
-import { cn } from "@/lib"
+import { cn } from "@/lib/utils"
 
 function DropdownMenu({
   ...props
