+++ react-app/src/components/ui/separator.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as SeparatorPrimitive from "@radix-ui/react-separator"
 
-import { cn } from "@/lib"
+import { cn } from "@/lib/utils"
 
 function Separator({
   className,
