+++ react-app/src/components/ui/tooltip.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as TooltipPrimitive from "@radix-ui/react-tooltip"
 
-import { cn } from "@/lib"
+import { cn } from "@/lib/utils"
 
 function TooltipProvider({
   delayDuration = 0,
